moving qualityCheck to checkers folder
