+++ .gitignore
@@ -138,6 +138,7 @@ celerybeat.pid
 .env
 .envrc
 .venv
+myenv
 env/
 venv/
 ENV/
