--- TESTS_SUMMARY.md
@@ -1,131 +1,0 @@
-# Test Suite Summary - LLM Quality Checks
-
-## Test Coverage
-
-**Total: 92 tests passing, 4 skipped (optional dependencies)**
-
-### New Test Files Created
-
-1. **`tests/test_llm_backends.py`** - 22 tests
-   - Backend abstraction testing
-   - OpenAI integration (with mocking)
-   - Ollama integration (skipped - requires package)
-   - JSON parsing and error handling
-   - Factory function testing
-
-2. **`tests/test_prompts.py`** - 14 tests
-   - All prompt template functions
-   - JSON structure validation
-   - Example request verification
-   - Severity level specification
-   - API name inclusion
-
-3. **`tests/test_llm_checker.py`** - 11 tests
-   - QualityChecker initialization
-   - Single API quality checks
-   - Module-level quality checks
-   - Sampling functionality
-   - Error handling
-   - Verbose output
-
-4. **`tests/test_integration.py`** - 7 tests
-   - End-to-end workflows
-   - Report formatting (text + JSON)
-   - Quality checks integration
-   - Multiple module support
-   - Error detection
-
-5. **`tests/test_checkers.py`** - Extended with 6 new tests
-   - Quality check integration in DriftDetector
-   - Graceful dependency handling
-   - Sample rate functionality
-   - Backend error handling
-
-## Test Categories
-
-### Unit Tests (66 tests)
-- Individual component testing
-- Mocked dependencies
-- Edge case coverage
-- Error scenarios
-
-### Integration Tests (26 tests)
-- Full workflow testing
-- Component interaction
-- Real project structures
-- End-to-end validation
-
-## Coverage Highlights
-
-✅ **LLM Backends**
-- OpenAI client initialization & API calls
-- Error handling (missing packages, no API key)
-- JSON response parsing
-- Backend factory pattern
-
-✅ **Prompt Generation**
-- All 4 prompt types tested
-- JSON structure verification
-- Example inclusion validation
-- Context preservation
-
-✅ **Quality Checker**
-- API-level checks
-- Module-level checks
-- Sampling strategies
-- LLM failure handling
-
-✅ **Integration**
-- DriftDetector integration
-- CLI-style workflows
-- Report formatting
-- JSON serialization
-
-## Running Tests
-
-```bash
-# All tests
-pytest tests/ -v
-
-# Just new LLM tests
-pytest tests/test_llm_* tests/test_prompts.py tests/test_integration.py -v
-
-# With coverage
-pytest tests/ --cov=doc_checker --cov-report=term-missing
-
-# Skip slow tests (if any)
-pytest tests/ -m "not slow"
-```
-
-## Skipped Tests
-
-4 tests skipped (require optional `ollama` package):
-- `test_ollama_backend_init`
-- `test_ollama_backend_default_model`
-- `test_ollama_backend_generate`
-- `test_ollama_backend_service_not_running`
-
-**Reason:** These test actual Ollama integration which requires the package installed.
-They're tested via integration tests with mocking instead.
-
-## Test Quality
-
-- **Mocking:** Extensive use of mocks for external dependencies
-- **Fixtures:** Reusable test fixtures in conftest.py
-- **Error Cases:** Comprehensive error handling tests
-- **Real Scenarios:** Integration tests use realistic project structures
-
-## CI/CD Ready
-
-All tests:
-- Run quickly (<1 second total)
-- No external dependencies required
-- Work without network access
-- Compatible with pytest-cov for coverage reports
-
-## Future Improvements
-
-- Add performance benchmarks for LLM calls
-- Test actual Ollama integration (requires ollama service)
-- Add mutation testing for prompt templates
-- Test concurrent quality checks
