+++ src/doc_checker/checkers_folder/quality.py
@@ -2,7 +2,10 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from doc_checker.models import DriftReport
+from doc_checker.llm_backends import get_backend
+from doc_checker.models import DriftReport, QualityIssue
+from doc_checker.prompts import get_combined_quality_prompt
+from doc_checker.utils.code_analyzer import CodeAnalyzer
 
 from .base import Checker
 
@@ -10,8 +13,8 @@ from .base import Checker
 class LLMQualityChecker(Checker):
     """LLM-based docstring quality analysis.
 
-    Lazily imports QualityChecker from llm_checker in check() to avoid
-    hard deps on ollama/openai. Supports ollama and openai backends.
+    Supports ollama and openai backends. Checks docstring completeness,
+    clarity, and accuracy for public APIs.
     """
 
     def __init__(
@@ -37,13 +40,6 @@ class LLMQualityChecker(Checker):
     def check(self, report: DriftReport) -> None:
         """Run LLM quality checks; skip with warning if deps missing."""
         try:
-            from doc_checker.llm_checker import QualityChecker
-        except ImportError as e:
-            report.warnings.append(
-                f"Quality checks skipped: {e}. Install with: pip install doc-checker[llm]"
-            )
-            return
-        try:
             checker = QualityChecker(
                 self.root_path,
                 self.backend_type,
@@ -60,3 +56,170 @@ class LLMQualityChecker(Checker):
             report.quality_issues.extend(
                 checker.check_module_quality(module, self.verbose, self.sample_rate)
             )
+
+
+class QualityChecker:
+    """Check documentation quality using LLMs."""
+
+    def __init__(
+        self,
+        root_path: Path,
+        backend_type: str = "ollama",
+        model: str | None = None,
+        api_key: str | None = None,
+        ignore_submodules: set[str] | None = None,
+    ):
+        """Initialize quality checker.
+
+        Args:
+            root_path: Project root path
+            backend_type: "ollama" (default) or "openai"
+            model: Model name (uses defaults if None)
+            api_key: API key for cloud backends
+            ignore_submodules: Submodule names to skip.
+
+        Raises:
+            ImportError: If backend package not installed
+            RuntimeError: If backend not available
+        """
+        self.root_path = root_path
+        self.code_analyzer = CodeAnalyzer(root_path)
+        self.backend = get_backend(backend_type, model, api_key)
+        self.ignore_submodules = ignore_submodules
+
+    def check_api_quality(
+        self, api_name: str, module_name: str, verbose: bool = False
+    ) -> list[QualityIssue]:
+        """Check quality of single API documentation.
+
+        Args:
+            api_name: API name (e.g., "MPS.evolve")
+            module_name: Module name (e.g., "emu_mps")
+            verbose: Print progress
+
+        Returns:
+            List of quality issues found
+        """
+        # Get API info
+        apis = self.code_analyzer.get_public_apis(module_name)
+        api_info = next((api for api in apis if api.name == api_name), None)
+
+        if not api_info:
+            return [
+                QualityIssue(
+                    api_name=f"{module_name}.{api_name}",
+                    severity="critical",
+                    category="error",
+                    message=f"API {api_name} not found in module {module_name}",
+                    suggestion="Check API name spelling",
+                    line_reference=None,
+                )
+            ]
+
+        if not api_info.docstring:
+            return [
+                QualityIssue(
+                    api_name=f"{module_name}.{api_name}",
+                    severity="critical",
+                    category="completeness",
+                    message="No docstring found",
+                    suggestion="Add docstring explaining what this API does",
+                    line_reference=None,
+                )
+            ]
+
+        # Build signature string
+        params_str = ", ".join(api_info.parameters)
+        return_str = (
+            f" -> {api_info.return_annotation}" if api_info.return_annotation else ""
+        )
+        signature = f"def {api_name}({params_str}){return_str}"
+
+        if verbose:
+            print(f"  Checking {module_name}.{api_name}...")
+
+        # Get LLM evaluation
+        prompt = get_combined_quality_prompt(
+            signature=signature,
+            docstring=api_info.docstring,
+            api_name=f"{module_name}.{api_name}",
+        )
+
+        try:
+            response = self.backend.generate_json(prompt)
+        except Exception as e:
+            return [
+                QualityIssue(
+                    api_name=f"{module_name}.{api_name}",
+                    severity="warning",
+                    category="error",
+                    message=f"LLM check failed: {e}",
+                    suggestion="Check LLM backend connection",
+                    line_reference=None,
+                )
+            ]
+
+        # Parse response
+        issues = []
+        for issue_data in response.get("issues", []):
+            issues.append(
+                QualityIssue(
+                    api_name=f"{module_name}.{api_name}",
+                    severity=issue_data.get("severity", "warning"),
+                    category=issue_data.get("category", "unknown"),
+                    message=issue_data.get("message", "No message"),
+                    suggestion=issue_data.get("suggestion", "No suggestion"),
+                    line_reference=issue_data.get("line_reference"),
+                )
+            )
+
+        if verbose and issues:
+            print(f"    Found {len(issues)} issues (score: {response.get('score', 0)})")
+
+        return issues
+
+    def check_module_quality(
+        self, module_name: str, verbose: bool = False, sample_rate: float = 1.0
+    ) -> list[QualityIssue]:
+        """Check quality of all APIs in a module.
+
+        Args:
+            module_name: Module to check (e.g., "emu_mps")
+            verbose: Print progress
+            sample_rate: Check only this fraction of APIs (0.0-1.0)
+
+        Returns:
+            List of all quality issues found
+        """
+        import random
+
+        apis, _ = self.code_analyzer.get_all_public_apis(
+            module_name, self.ignore_submodules
+        )
+
+        if not apis:
+            if verbose:
+                print(f"No public APIs found in {module_name}")
+            return [
+                QualityIssue(
+                    api_name=module_name,
+                    severity="warning",
+                    category="error",
+                    message=f"No public APIs found in module {module_name}",
+                    suggestion="Check module name or ensure it is installed",
+                    line_reference=None,
+                )
+            ]
+
+        if sample_rate < 1.0:
+            apis = random.sample(apis, int(len(apis) * sample_rate))
+
+        if verbose:
+            print(f"Checking {len(apis)} APIs in {module_name}...")
+
+        all_issues = []
+        for api in apis:
+            issues = self.check_api_quality(api.name, module_name, verbose)
+            all_issues.extend(issues)
+
+        return all_issues
